requirements.txt
async-timeout==4.0.3
cx-Oracle==8.3.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0
redis==5.0.4
six==1.16.0
tzdata==2024.1
avro==1.11.3
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.4.0
fastavro==1.9.4
idna==3.7
requests==2.32.3
urllib3==2.2.1
elastic-transport==8.13.1
elasticsearch==8.13.2
